Updateclient.py
New Client, It still connects to the current server, it also creates a dictionary based on inputs. Currently the dictionary is for our group or other groups, When you enter a value for num_members it represents the number of group members you want a dictionary created for, the after that it will ask for inputs based on fixed criteria, name, surname and project role. The code then creates a list of containing multiple dictionaries for each group member.
This list is subsequently written to a binary and JSON file based on your needs.
I have the old code copy saved for control purposes